Moar eigen Eigenvalues.cpp: Deleted!
git-svn-id: http://theory.cm.utexas.edu/svn/eon@499 9d962951-508e-48a8-94b5-974f9cfda47c

+++ client_eigen/potentials/QSC/QSC.h
@@ -12,12 +12,12 @@ class QSC
         ~QSC(void);
         // To satify interface
         void initialize();
-        void initialize(long N, const double *R, const long *atomicNrs,
+        void initialize(long N, const double *R, const int *atomicNrs,
                         const double *box);
         void cleanMemory();
-        void force(long N, const double *R, const long *atomicNrs,
+        void force(long N, const double *R, const int *atomicNrs,
                    double *F, double *U, const double *box);
-        void energy(long N, const double *R, const long *atomicNrs,
+        void energy(long N, const double *R, const int *atomicNrs,
                     double *U, const double *box);
     private:
         bool init;
